topnav is gone for !user and logo redirects to user-homepage
